FIX-2121 move hook
