+++ Src/WitsmlExplorer.Frontend/components/Modals/MissingDataAgentModal.tsx
@@ -6,13 +6,14 @@ import { v4 as uuid } from "uuid";
 import OperationContext from "../../contexts/operationContext";
 import OperationType from "../../contexts/operationType";
 import useExport from "../../hooks/useExport";
+import { useLocalStorageState } from "../../hooks/useLocalStorageState";
 import MissingDataJob, { MissingDataCheck } from "../../models/jobs/missingDataJob";
 import WellReference from "../../models/jobs/wellReference";
 import WellboreReference from "../../models/jobs/wellboreReference";
 import { ObjectType } from "../../models/objectType";
 import JobService, { JobType } from "../../services/jobService";
 import { Colors } from "../../styles/Colors";
-import { STORAGE_MISSING_DATA_AGENT_CHECKS_KEY, useLocalStorageState } from "../../tools/localStorageHelpers";
+import { STORAGE_MISSING_DATA_AGENT_CHECKS_KEY } from "../../tools/localStorageHelpers";
 import { StyledAccordionHeader } from "./LogComparisonModal";
 import { objectToProperties, selectAllProperties } from "./MissingDataAgentProperties";
 import ModalDialog, { ModalContentLayout, ModalWidth } from "./ModalDialog";
